+++ resolwe/flow/serializers/entity.py
@@ -1,26 +1,26 @@
 """Resolwe entity serializer."""
 from resolwe.flow.models import Collection, Entity
 
-from .collection import CollectionSerializer
+from .collection import BaseCollectionSerializer
 from .fields import DictRelatedField
 
 
-class EntitySerializer(CollectionSerializer):
+class EntitySerializer(BaseCollectionSerializer):
     """Serializer for Entity."""
 
     collection = DictRelatedField(
         queryset=Collection.objects.all(),
-        serializer=CollectionSerializer,
+        serializer=BaseCollectionSerializer,
         allow_null=True,
         required=False,
         write_permission="edit",
     )
 
-    class Meta(CollectionSerializer.Meta):
+    class Meta(BaseCollectionSerializer.Meta):
         """EntitySerializer Meta options."""
 
         model = Entity
-        fields = CollectionSerializer.Meta.fields + (
+        fields = BaseCollectionSerializer.Meta.fields + (
             "collection",
             "duplicated",
             "type",